backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
httpx==0.24.1
pydantic==2.4.2
python-dotenv==1.0.0
sqlalchemy==2.0.20
aiohttp==3.8.5
pytest==8.3.2
flake8==7.1.0
python-jose==3.3.0
bcrypt==4.2.0